actual

+++ src/controllers/usuarioController.ts
@@ -37,6 +37,9 @@ class UsuarioController {
         if (!req.body.password) {
             return res.status(400).json({ message: "La contraseña no puede estar vacía" });
         }
+        if (!req.body.role) {
+            return res.status(400).json({ message: "El rol no puede estar vacío" });
+        }
 
         const encryptedText = await utils.hashPassword(req.body.password);
         if (!encryptedText) {
@@ -71,7 +74,9 @@ public async update(req: Request, res: Response): Promise<void> {
           if (req.body.password && !validator.isStrongPassword(req.body.password)) {
             return res.status(400).json({ message: "La contraseña debe ser segura" });
           }
-
+          if (!req.body.role) {
+            return res.status(400).json({ message: "El rol no puede estar vacío" });
+        }
         if (req.body.usuario && req.body.usuario.password) {
             var encryptedText = await utils.hashPassword(req.body.password);
             req.body.password = encryptedText;
@@ -90,27 +95,39 @@ public async update(req: Request, res: Response): Promise<void> {
 }
 
 
-    public async delete(req: Request, res: Response): Promise<any> {
-        try {
-            const { email } = req.body;
+public async delete(req: Request, res: Response): Promise<any> {
+    try {
+        const { email } = req.body;
 
-            console.log('Email:', email);
-
-            if (!validator.isEmail(email)) {
-                return res.status(400).json({ message: "Correo electrónico no válido" });
-            }
-
-            const connection = await pool;
-            const result = await connection.query(
-                'DELETE FROM tbl_usuario WHERE email = ?',
-                [email]
-            );
-
-            return res.json({ text: "Usuario con el correo " + email + " ha sido eliminado" });
-        } catch (error: any) {
-            return res.status(500).json({ message: `${error.message}` });
+        
+        if (!validator.isEmail(email)) {
+            return res.status(400).json({ message: "Correo electrónico no válido" });
         }
+
+       
+        const connection = await pool;
+        const userExists = await connection.query(
+            'SELECT * FROM tbl_usuario WHERE email = ?',
+            [email]
+        );
+
+       
+        if (userExists.length === 0) {
+            return res.status(404).json({ message: "El usuario con este correo electrónico no existe" });
+        }
+
+       
+        const result = await connection.query(
+            'DELETE FROM tbl_usuario WHERE email = ?',
+            [email]
+        );
+
+        return res.json({ text: "Usuario con el correo " + email + " ha sido eliminado" });
+    } catch (error: any) {
+        return res.status(500).json({ message: `${error.message}` });
     }
+}
+
 }
 export const usuarioController = new UsuarioController();
